Fix `add_user()` in `user.py`
After running the function to add a user
to a workspace in Clockify via Timekeeper, I hit a
400 error. There was something wrong with the
payload.

So I looked into the `add_user()` function in this
client. It appeared outdated; the payload format
was incorrect. This should take a dict in this
format:
{
    "email": "example@example.com"
}
But the old function was sending the payload in
the following format instead:
{
    "emails": ["example@example.com"]
}
I updated this and now everything works as
expected.